chore: update evaluation
