+++ packages/evaluation/page-cases/inspect/online_order_list.json
@@ -1,36 +1,45 @@
 {
-    "testDataPath": "online_order_list",
-    "testCases": [{
-            "prompt": "'清爽不喝腻'下面第二个饮品的名称",
-            "multi": false,
-            "response": [{
-                "id": "eabha",
-                "indexId": 20
-            }]
-        },
+  "testDataPath": "online_order_list",
+  "testCases": [
+    {
+      "prompt": "'清爽不喝腻'下面第二个饮品的名称",
+      "multi": false,
+      "response": [
         {
-            "prompt": "多肉葡萄的选择规格按钮",
-            "multi": false,
-            "response": [{
-                "id": "kmdfd",
-                "indexId": 39
-            }]
-        },
-        {
-            "prompt": "多肉葡萄的价格",
-            "multi": false,
-            "response": [{
-                "id": "pceib",
-                "indexId": 38
-            }]
-        },
-        {
-            "prompt": "左侧导航栏’要雪糕‘",
-            "multi": false,
-            "response": [{
-                "id": "lalae",
-                "indexId": 9
-            }]
+          "id": "eabha",
+          "indexId": 20
         }
-    ]
+      ]
+    },
+    {
+      "prompt": "多肉葡萄的选择规格按钮",
+      "multi": false,
+      "response": [
+        {
+          "id": "kmdfd",
+          "indexId": 39
+        }
+      ]
+    },
+    {
+      "prompt": "多肉葡萄的价格",
+      "multi": false,
+      "response": [
+        {
+          "id": "pceib",
+          "indexId": 38
+        }
+      ]
+    },
+    {
+      "prompt": "左侧导航栏’要雪糕‘",
+      "multi": false,
+      "response": [
+        {
+          "id": "lalae",
+          "indexId": 9
+        }
+      ]
+    }
+  ]
 }
