+++ packages/midscene/package.json
@@ -8,21 +8,39 @@
   "type": "commonjs",
   "main": "./dist/lib/index.js",
   "types": "./dist/lib/types/index.d.ts",
-  "files": ["dist", "report", "README.md"],
+  "files": [
+    "dist",
+    "report",
+    "README.md"
+  ],
   "exports": {
     ".": "./dist/lib/index.js",
     "./env": "./dist/lib/env.js",
     "./utils": "./dist/lib/utils.js",
     "./ai-model": "./dist/lib/ai-model.js",
-    "./tree": "./dist/lib/tree.js"
+    "./tree": "./dist/lib/tree.js",
+    "./evaluation": "./dist/lib/evaluation.js"
   },
   "typesVersions": {
     "*": {
-      ".": ["./dist/lib/types/index.d.ts"],
-      "env": ["./dist/lib/types/env.d.ts"],
-      "utils": ["./dist/lib/types/utils.d.ts"],
-      "ai-model": ["./dist/lib/types/ai-model.d.ts"],
-      "tree": ["./dist/lib/types/tree.d.ts"]
+      ".": [
+        "./dist/lib/types/index.d.ts"
+      ],
+      "env": [
+        "./dist/lib/types/env.d.ts"
+      ],
+      "utils": [
+        "./dist/lib/types/utils.d.ts"
+      ],
+      "ai-model": [
+        "./dist/lib/types/ai-model.d.ts"
+      ],
+      "tree": [
+        "./dist/lib/types/tree.d.ts"
+      ],
+      "evaluation": [
+        "./dist/lib/types/evaluation.d.ts"
+      ]
     }
   },
   "scripts": {
@@ -35,11 +53,6 @@
     "test:ai": "AITEST=true npm run test",
     "computer": "TEST_COMPUTER=true npm run test:ai -- tests/ai/evaluate/computer.test.ts",
     "test:parse-action": "npm run test:ai -- tests/ai/parse-action.test.ts",
-    "evaluate": "npm run test:ai -- tests/ai/evaluate/inspect.test.ts",
-    "evaluate:assertion": "npm run test:ai -- tests/ai/evaluate/assertion.test.ts",
-    "evaluate:plan": "npm run test:ai -- tests/ai/evaluate/plan/planning.test.ts",
-    "evaluate:update": "UPDATE_AI_DATA=true npm run test:ai -- tests/ai/evaluate/inspect.test.ts",
-    "prompt": "npm run test:ai -- tests/ai/parse-action.test.ts",
     "prepublishOnly": "npm run build"
   },
   "dependencies": {
