+++ packages/midscene/tests/ai/extract/extract.test.ts
@@ -1,6 +1,6 @@
 import { AiExtractElementInfo } from '@/ai-model';
+import { getContextFromFixture } from '@/evaluation';
 import { describe, expect, it, vi } from 'vitest';
-import { getPageDataOfTestName } from '../evaluate/test-suite/util';
 
 vi.setConfig({
   testTimeout: 180 * 1000,
@@ -9,7 +9,7 @@ vi.setConfig({
 
 describe('extract', () => {
   it('todo', async () => {
-    const { context } = await getPageDataOfTestName('todo');
+    const { context } = await getContextFromFixture('todo');
 
     const { parseResult } = await AiExtractElementInfo({
       dataQuery: 'Array<string>, Complete task list, string is the task',
@@ -19,7 +19,7 @@ describe('extract', () => {
   });
 
   it('online order', async () => {
-    const { context } = await getPageDataOfTestName('online_order');
+    const { context } = await getContextFromFixture('online_order');
 
     const { parseResult } = await AiExtractElementInfo({
       dataQuery: '{name: string, price: string}[], 饮品名称和价格',
@@ -29,7 +29,7 @@ describe('extract', () => {
   });
 
   it('todo obj', async () => {
-    const { context } = await getPageDataOfTestName('todo');
+    const { context } = await getContextFromFixture('todo');
 
     const { parseResult } = await AiExtractElementInfo({
       dataQuery:
