+++ packages/midscene/tests/ai/ui-tars-planning/output.png
