+++ packages/midscene/vitest.config.ts
@@ -22,9 +22,6 @@ export default defineConfig({
     include: enableAiTest
       ? ['tests/ai/**/**.test.ts', ...basicTest]
       : basicTest,
-    exclude: process.env.TEST_COMPUTER
-      ? []
-      : ['tests/ai/evaluate/computer.test.ts'],
   },
   define: {
     __VERSION__: `'${version}'`,
